Adds appDelegate h&m files, updates deprecated code, prunes commented-out sections, adds helloController h&m files

+++ UIColor-Expanded.m
@@ -398,7 +398,7 @@ static NSLock *colorNameCacheLock;
 }
 
 - (NSString *)hexStringFromColor {
-	return [NSString stringWithFormat:@"%0.6X", self.rgbHex];
+	return [NSString stringWithFormat:@"%0.6X", (unsigned int)self.rgbHex];
 }
 
 - (NSString *)closestColorName {
